moved StateFileConfig to pipeline config
